# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk platform ujian online IPPAT (Fiber + GORM + PostgreSQL). Semua response dibungkus format {success, message, data, errors}.\nEndpoint yang butuh login (guru/admin/peserta) ditandai kunci gembok di UI ini — klik \"Authorize\" dan isi `Bearer \u003ctoken\u003e` yang didapat dari /api/auth/login atau /api/auth/register.",
        "title": "IPPAT Backend API",
        "contact": {
            "name": "IPPAT Backend Team"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/soal": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soal"
                ],
                "summary": "List soal",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Nomor halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah data per halaman",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_soal_dto.SoalListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Form-data (bukan JSON) karena mendukung upload gambar untuk soal maupun tiap opsi jawaban.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soal"
                ],
                "summary": "Buat soal baru",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Bank Soal",
                        "name": "id_bank_soal",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Nomor urut soal",
                        "name": "no_soal",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks soal",
                        "name": "soal",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Gambar soal",
                        "name": "gambar_soal",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks opsi A",
                        "name": "opsi_a",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks opsi B",
                        "name": "opsi_b",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks opsi C",
                        "name": "opsi_c",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks opsi D",
                        "name": "opsi_d",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Teks opsi E",
                        "name": "opsi_e",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Gambar opsi A",
                        "name": "gambar_a",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Gambar opsi B",
                        "name": "gambar_b",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Gambar opsi C",
                        "name": "gambar_c",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Gambar opsi D",
                        "name": "gambar_d",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Gambar opsi E",
                        "name": "gambar_e",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Kunci jawaban (A/B/C/D/E)",
                        "name": "kunci",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_soal_dto.SoalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/soal/bank/{bank_soal_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soal"
                ],
                "summary": "List soal berdasarkan bank soal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Bank Soal",
                        "name": "bank_soal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Nomor halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah data per halaman",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_soal_dto.SoalListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/soal/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload file .xls/.xlsx (maks 10MB) berisi banyak soal sekaligus untuk satu bank soal. Response berisi ringkasan jumlah berhasil/gagal beserta detail error per baris.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soal"
                ],
                "summary": "Import soal massal dari file Excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Bank Soal",
                        "name": "id_bank_soal",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File Excel (.xls/.xlsx, maks 10MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_soal_dto.ImportSoalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
...
        "/soal/{id}/restore": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soal"
                ],
                "summary": "Pulihkan soal yang sudah dihapus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Soal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend_internal_helpers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "backend_internal_modules_soal_dto.ImportSoalErrorDetail": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "row": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_modules_soal_dto.ImportSoalResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_soal_dto.ImportSoalErrorDetail"
                    }
                },
                "import_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "timestamp": {
                    "type": "string"
                },
                "total_failed": {
                    "type": "integer"
                },
                "total_processed": {
                    "type": "integer"
                },
                "total_success": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_modules_soal_dto.SoalListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_soal_dto.SoalResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_modules_soal_dto.SoalResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "gambar_a": {
                    "type": "string"
                },
                "gambar_b": {
                    "type": "string"
                },
                "gambar_c": {
                    "type": "string"
                },
                "gambar_d": {
                    "type": "string"
                },
                "gambar_e": {
                    "type": "string"
                },
                "gambar_soal": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "id_bank_soal": {
                    "type": "string"
                },
                "kunci": {
                    "type": "string"
                },
                "no_soal": {
                    "type": "integer"
                },
                "opsi_a": {
                    "type": "string"
                },
                "opsi_b": {
                    "type": "string"
                },
                "opsi_c": {
                    "type": "string"
                },
                "opsi_d": {
                    "type": "string"
                },
                "opsi_e": {
                    "type": "string"
                },
                "soal": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Ketik \"Bearer\" diikuti spasi dan JWT token. Contoh: \"Bearer eyJhbGciOi...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}